requirements.txt
fastapi
uvicorn
python-multipart
spacy
python-docx
PyPDF2
